backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "PitchPerfect"
version = "0.1.0"
description = "AI-powered pitch analysis and feedback system"
readme = "README.md"
authors = [{name = "PitchPerfect Team"}]
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "fastapi>=0.95.0",
    "uvicorn>=0.22.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai-whisper>=20231117",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "spacy>=3.5.0",
    "textstat>=0.7.3",
    "requests>=2.28.0",
    "httpx>=0.24.0",
    "pytest>=7.3.1",
    "pytest-asyncio>=0.21.0",
    "tqdm>=4.65.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/PitchPerfect"
Repository = "https://github.com/yourusername/PitchPerfect"

[tool.setuptools.packages.find]
where = ["backend"]
include = ["app*"]

[tool.setuptools.package-dir]
"" = "backend"
